Refactor common test helpers
